refactor: move out thunks
