+++ src/components/organisms/GitCloneModal/GitCloneModal.tsx
@@ -14,10 +14,10 @@ import styled from 'styled-components';
 
 import {DEFAULT_GIT_REPO_PLACEHOLDER, VALID_URL_REGEX} from '@constants/constants';
 
-import {setCreateProject} from '@redux/appConfig';
 import {closeGitCloneModal} from '@redux/git';
 import {cloneGitRepo} from '@redux/git/git.ipc';
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
+import {setCreateProject} from '@redux/thunks/project';
 
 import {FileExplorer} from '@atoms';
