+++ src/redux/reducers/ui.ts
@@ -7,13 +7,13 @@ import {Entries} from 'type-fest';
 
 import {DEFAULT_PANE_CONFIGURATION} from '@constants/constants';
 
-import {setOpenProject} from '@redux/appConfig';
 import {connectCluster} from '@redux/cluster/thunks/connect';
 import initialState from '@redux/initialState';
 import {previewSavedCommand} from '@redux/services/previewCommand';
 import {stopClusterConnection} from '@redux/thunks/cluster';
 import {previewHelmValuesFile} from '@redux/thunks/previewHelmValuesFile';
 import {previewKustomization} from '@redux/thunks/previewKustomization';
+import {setOpenProject} from '@redux/thunks/project';
 import {runPreviewConfiguration} from '@redux/thunks/runPreviewConfiguration';
 import {setRootFolder} from '@redux/thunks/setRootFolder';
