+++ src/redux/thunks/project/index.ts
@@ -1,0 +1,4 @@
+export {setCreateProject} from './createProject';
+export {setDeleteProject} from './deleteProject';
+export {setOpenProject} from './openProject';
+export {toggleEventTracking} from './toggleEventTracking';
